requirements.txt
# Database
psycopg2-binary>=2.9
sqlalchemy==2.0.23

# Data Processing
pandas==2.1.3
numpy==1.26.2

# API Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Validation
email-validator==2.1.0
